+++ instrumentation/opentelemetry-instrumentation-django/tests/test_sqlcommenter.py
@@ -13,15 +13,16 @@
 # limitations under the License.
 
 # pylint: disable=no-name-in-module
-
 from unittest.mock import MagicMock, patch
 
+import pytest
 from django import VERSION, conf
 from django.http import HttpResponse
 from django.test.utils import setup_test_environment, teardown_test_environment
 
 from opentelemetry.instrumentation.django import DjangoInstrumentor
 from opentelemetry.instrumentation.django.middleware.sqlcommenter_middleware import (
+    SqlCommenter,
     _QueryWrapper,
 )
 from opentelemetry.test.wsgitestutil import WsgiTestBase
@@ -98,3 +99,19 @@ class TestMiddleware(WsgiTestBase):
             "Select 1 /*app_name='app',controller='view',route='route',traceparent='%%2Atraceparent%%3D%%2700-0000000"
             "00000000000000000deadbeef-000000000000beef-00'*/",
         )
+
+    @patch(
+        "opentelemetry.instrumentation.django.middleware.sqlcommenter_middleware._QueryWrapper"
+    )
+    def test_multiple_connection_support(self, query_wrapper):
+        if not DJANGO_2_0:
+            pytest.skip()
+
+        requests_mock = MagicMock()
+        get_response = MagicMock()
+
+        sql_instance = SqlCommenter(get_response)
+        sql_instance(requests_mock)
+
+        # check if query_wrapper is added to the context for 2 databases
+        self.assertEqual(query_wrapper.call_count, 2)