+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.12.0rc2-0.32b0...HEAD)
+- Adding multiple db connections support for django-instrumentation's sqlcommenter
+  ([#1187](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1187))
 
 ### Added
 - `opentelemetry-instrumentation-redis` add support to instrument RedisCluster clients